move reviews response to correct folder
