+++ src/Mantasflowers.Services/Services/Review/IProductReviewService.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Mantasflowers.Contracts.Review;
+using Mantasflowers.Contracts.Review.Response;
 
 namespace Mantasflowers.Services.Services.Review
 {
